Some 64bit cleanups for Windows

+++ src/core/PWScore.cpp
@@ -970,7 +970,7 @@ static void ManageIncBackupFiles(const stringT &cs_filenamebase,
     while (file_nums[x++] == nnn && x < file_nums.size())
       nnn++;
     // Now we need to determine who to delete.
-    int next = 999 - (maxnumincbackups - nnn);
+    size_t next = 999 - (maxnumincbackups - nnn);
     int m = 1;
     for (x = 0; x < file_nums.size(); x++)
       if (file_nums[x] < next)
@@ -1510,7 +1510,7 @@ bool PWScore::Validate(const size_t iMAXCHARS, const bool bInReadfile,
                  bInReadfile ? _T("true") : _T("false"), pRpt);
 
   int n = -1;
-  unsigned int uimaxsize(0);
+  size_t uimaxsize(0);
 
   MultiCommands *pmulticmds(NULL);
